Intialize dependencies on creation (see #4)
We need to transfer the field dependencies to our JavaScript. So we have
to create a dependencies array from dca data containing all editable
fields and their dependent fields. If we use the same array in the
setFieldDependencies method (instead of all tl_member fields) we have a
much smaller array to walk through.